// src/lib/impakt-dev-api-client/react-query/types/getRoutineRes.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * Impakt API
 * OpenAPI spec version: 1.0.0
 */
import type { GetRoutineResType } from './getRoutineResType';
import type { GetUserRes } from './getUserRes';
import type { GetTimelineBlockRes } from './getTimelineBlockRes';
import type { TagRes } from './tagRes';

export interface GetRoutineRes {
  id: number;
  name: string;
  type: GetRoutineResType;
  /** estimated time in seconds */
  estimatedTime: number;
  creatorId?: number;
  /** available only with Get methods */
  Creator?: GetUserRes;
  /** available only with Get methods */
  TimelineBlocks?: GetTimelineBlockRes[];
  createdAt: string;
  updatedAt: string;
  tags: TagRes[];
  totalCount: number;
  /** Is routine duplicated for challenge */
  isChallengeRoutine: boolean;
}